rename to signed entity service
