+++ mithril-aggregator/src/artifact_builder/mod.rs
@@ -1,11 +1,8 @@
 //! The module used for building artifact
-
-mod artifact_builder_service;
 mod cardano_immutable_files_full;
 mod interface;
 mod mithril_stake_distribution;
 
-pub use artifact_builder_service::*;
 pub use cardano_immutable_files_full::*;
 pub use interface::*;
 pub use mithril_stake_distribution::*;
