+++ mithril-aggregator/src/signed_entity_service.rs
